* master: (42 commits)
  Fixed everything needed for windows integration
  Refactored command calls
  Testingfoo
  Refactored some stuff due to unit testing
  Refactored some stuff due to unit testing
  Refactored some stuff due to unit testing
  Refactored some stuff due to unit testing
  Refactored some stuff due to unit testing
  #135 Implemented fix for repo schenanigans
  #145 Fixed deletion
  #140 Implemented missing xmlrpc for php 7.4
  #141 Exposed proxy ports
  Increased sleep time in importer
  Increased upload limit in all php versions
  Current state
  #138 Rerouted doil base help
  Hotfix
  Hotfix
  Hotfix
  Hotfix
  ...

# Conflicts:
#	src/lib/instances/apply/apply.sh
#	src/lib/instances/create/create.sh
#	src/stack/states/php7.0/php/init.sls
#	src/stack/states/php7.1/php/init.sls
#	src/stack/states/php7.2/php/init.sls
#	src/stack/states/php7.3/php/init.sls
#	src/stack/states/php7.4/php/init.sls
#	src/stack/states/php8.0/php/init.sls
#	src/templates/minion/docker-compose.yml

+++ update.sh
@@ -105,44 +105,76 @@ fi
 
 printf " ${GREEN}ok${NC}\n"
 
-echo -n "Adding group doil ..."
-
-groupadd doil
-
-printf " ${GREEN}ok${NC}\n"
+CHECK_GROUP=$(grep doil /etc/group)
+if [[ -z ${CHECK_GROUP} ]]
+then
+  echo -n "Adding group doil ..."
+  groupadd doil
+  printf " ${GREEN}ok${NC}\n"
+fi
 
 # create the log file
-echo -n "Creating log file in /var/log/doil.log ..."
+if [ ! -f /var/log/doil.log ]
+then
+  echo -n "Creating log file in /var/log/doil.log ..."
 
-touch /var/log/doil.log
-chown root:doil /var/log/doil.log
-chmod 777 /var/log/doil.log
+  touch /var/log/doil.log
+  chown root:doil /var/log/doil.log
+  chmod 777 /var/log/doil.log
 
-printf " ${GREEN}ok${NC}\n"
+  printf " ${GREEN}ok${NC}\n"
+fi
 
 # create mandatory folders
 echo -n "Creating mandatory folders ..."
 
-mkdir /usr/local/lib/doil
-mkdir /usr/local/lib/doil/lib
-mkdir /usr/local/lib/doil/server
+if [ ! -d /usr/local/lib/doil ]
+then
+  mkdir /usr/local/lib/doil
+fi
+if [ ! -d /usr/local/lib/doil/lib ]
+then
+  mkdir /usr/local/lib/doil/lib
+fi
+if [ ! -d /usr/local/lib/doil/server ]
+then
+  mkdir /usr/local/lib/doil/server
+fi
 chown -R root:doil /usr/local/lib/doil
 
-mkdir /etc/doil/
-chown -R root:doil /etc/doil
-chmod -R g+w /etc/doil
-chmod -R g+s /etc/doil
+if [ ! -d /etc/doil ]
+then
+  mkdir /etc/doil/
+  chown -R root:doil /etc/doil
+  chmod -R g+w /etc/doil
+  chmod -R g+s /etc/doil
+fi
 
-mkdir /usr/local/share/doil
-mkdir /usr/local/share/doil/templates
-mkdir /usr/local/share/doil/stack
-mkdir /usr/local/share/doil/instances
-mkdir /usr/local/share/doil/repositories
-chown -R root:doil /usr/local/share/doil
-chmod -R g+w /usr/local/share/doil/repositories
-chmod -R g+w /usr/local/share/doil/instances
-chmod -R g+s /usr/local/share/doil/repositories
-chmod -R g+s /usr/local/share/doil/instances
+if [ ! -d /usr/local/share/doil ]
+then
+  mkdir /usr/local/share/doil
+  chown -R root:doil /usr/local/share/doil
+fi
+if [ ! -d /usr/local/share/doil/templates ]
+then
+  mkdir /usr/local/share/doil/templates
+fi
+if [ ! -d /usr/local/share/doil/instances ]
+then
+  mkdir /usr/local/share/doil/instances
+fi
+if [ ! -d /usr/local/share/doil/instances ]
+then
+  mkdir /usr/local/share/doil/instances
+  chmod -R g+w /usr/local/share/doil/instances
+  chmod -R g+s /usr/local/share/doil/instances
+fi
+if [ ! -d /usr/local/share/doil/repositories ]
+then
+  mkdir /usr/local/share/doil/repositories
+  chmod -R g+w /usr/local/share/doil/repositories
+  chmod -R g+s /usr/local/share/doil/repositories
+fi
 
 printf " ${GREEN}ok${NC}\n"
 
@@ -159,8 +191,7 @@ chmod -R +x /usr/local/lib/doil/lib
 
 cp -r src/server/* /usr/local/lib/doil/server/
 chown -R root:doil /usr/local/lib/doil/server
-chmod -R g+w /usr/local/lib/doil/server/proxy/conf/sites
-chmod -R g+s /usr/local/lib/doil/server/proxy/conf/sites
+chmod -R 777 /usr/local/lib/doil/server/proxy/conf/
 
 cp -r src/templates/* /usr/local/share/doil/templates
 chown root:doil /usr/local/share/doil/templates
@@ -173,13 +204,23 @@ printf " ${GREEN}ok${NC}\n"
 # setting up basic configuration
 echo -n "Setting up basic configuration ..."
 
-cp src/conf/doil.conf /etc/doil/doil.conf
-touch /etc/doil/repositories.conf
-chown -R root:doil /etc/doil/
-chmod g+w /etc/doil/repositories.conf
-touch /etc/doil/user.conf
+if [ ! -f /etc/doil/doil.conf ]
+then
+  cp src/conf/doil.conf /etc/doil/doil.conf
+fi
 
-echo "ilias=git@github.com:ILIAS-eLearning/ILIAS.git" > "/etc/doil/repositories.conf"
+if [ ! -f /etc/doil/repositories.conf ]
+then
+  touch /etc/doil/repositories.conf
+  chmod g+w /etc/doil/repositories.conf
+fi
+
+if [ ! -f /etc/doil/user.conf ]
+then
+  touch /etc/doil/user.conf
+  chmod g+w /etc/doil/user.conf
+fi
+chown -R root:doil /etc/doil/
 
 IPEXIST=$(grep "172.24.0.254" /etc/hosts)
 if [[ -z ${IPEXIST} ]]
@@ -192,24 +233,38 @@ printf " ${GREEN}ok${NC}\n"
 echo -n "Move userdata ..."
 
 HOME=$(eval echo "~${SUDO_USER}")
-mv ${HOME}/.doil/config/repos ${HOME}/.doil/config/repositories.conf
+if [ ! -f ${HOME}/.doil/config/repositories.conf ]
+then
+  mv ${HOME}/.doil/config/repos ${HOME}/.doil/config/repositories.conf
+fi
 
-mkdir ${HOME}/.doil/repositories
-mv /usr/local/lib/doil/tpl/repos ${HOME}/.doil/repositories
+if [ ! -d ${HOME}/.doil/repositories ]
+then
+  mkdir ${HOME}/.doil/repositories
+  mv /usr/local/lib/doil/tpl/repos ${HOME}/.doil/repositories
+fi
 
-mkdir ${HOME}/.doil/instances
-for LINK in $(ls ${HOME}/.doil)
-do
-  if [[ ${LINK} == "config" ]]
-  then
-    continue
-  fi
-  mv ${LINK} ${HOME}/.doil/instances/${LINK}
-done
+if [ ! -d ${HOME}/.doil/instances ]
+then
+  mkdir ${HOME}/.doil/instances
+  for LINK in $(ls ${HOME}/.doil)
+  do
+    if [[ ${LINK} == "config" ]]
+    then
+      continue
+    fi
+    mv ${LINK} ${HOME}/.doil/instances/${LINK}
+  done
+fi
 
 chown -R ${SUDO_USER}:${SODU_USER} "${HOME}/.doil"
 usermod -a -G doil ${SUDO_USER}
-echo "${SUDO_USER}">>"/etc/doil/user.conf"
+
+USEREXISTS=$(grep "${SUDO_USER}" /etc/doil/user.conf)
+if [[ -z ${USEREXISTS} ]]
+then
+  echo "${SUDO_USER}">>"/etc/doil/user.conf"
+fi
 
 printf " ${GREEN}ok${NC}\n"
 
@@ -221,8 +276,8 @@ printf " ${GREEN}ok${NC}\n"
 
 echo -n "Restarting server ..."
 
-doil system:proxy restart --quiet
-doil system:salt restart --quiet
+#doil system:proxy restart --quiet
+#doil system:salt restart --quiet
 
 printf " ${GREEN}ok${NC}\n"
